+++ src/error/display.rs
@@ -1,8 +1,8 @@
-use super::NetAddrError;
-use core::fmt::{Display, Error, Formatter};
+use super::Error;
+use core::fmt::{self, Display, Formatter};
 
-impl Display for NetAddrError {
-	fn fmt(&self, f: &mut Formatter) -> Result<(), Error> {
+impl Display for Error {
+	fn fmt(&self, f: &mut Formatter) -> Result<(), fmt::Error> {
 		match self {
 			Self::ParseError(text) => write!(f, "unable to parse address: {}", text),
 		}
@@ -15,7 +15,7 @@ mod tests {
 
 	#[test]
 	fn right_message() {
-		let error: NetAddrError = NetAddrError::ParseError("INNER_TEXT".into());
+		let error: Error = Error::ParseError("INNER_TEXT".into());
 		let result: &str = &format!("{}", error);
 		assert_eq!(result, "unable to parse address: INNER_TEXT");
 	}