+++ src/error/from/addr_parse_error.rs
@@ -1,7 +1,7 @@
-use super::NetAddrError;
+use super::Error;
 use std::net::AddrParseError;
 
-impl From<AddrParseError> for NetAddrError {
+impl From<AddrParseError> for Error {
 	fn from(other: AddrParseError) -> Self {
 		Self::ParseError(other.to_string())
 	}