+++ src/error/result.rs
@@ -1,0 +1,3 @@
+use super::Error;
+
+pub type Result<T> = core::result::Result<T, Error>;