+++ src/netaddr_error/from.rs
@@ -1,3 +1,0 @@
-use super::NetAddrError;
-
-mod addr_parse_error;